EF-212: Allow exception suppression when entity type property has ColumnAttribute
Currently, the provider throws when a ColumnAttribute is encountered. This can be helpful for people who think it will be used, when it isn't. However, there are cases where already existing entity types must be used, and in these cases the column attributes can be ignored. We can allow this by making this message into an exception by default, but changeable to a log or ignore, using EF's infrastructure.