EF-27: Start using the EF Core provider specification tests (#212)
* EF-27: Start using the EF Core provider specification tests

This PR starts using the published EF Core provider tests to get coverage for lots of cases. There are expected to be some tests that fail, for things not (yet) supported by the MongoDB provider. All failing tests have a "Fails: AV" comment. Damien and I will need to go through these and create issues for anything that needs it.

In addition other TODOs are:
- ColumnAttribute always throws
- Submit PR to EF Core so we can use their copy of the logger factory
- File an issue to not throw for keyless entity types

* Move specification tests into separate project

* Fix build breaks.

* Investigated failures

* EF-214: Allow keyless types to be used

* Updated with EF issue references

* Update evergreen to use .NET 9 SDK

* Reduce test output and evergreen tweak

* Evergreen

* More evergreen

* Test tweaks

* Review comments

* Install .NET 8 runtime instead of SDK. Leaves only the .NET 9 SDK installed.

* Try again